requirements.txt
Flask==2.3.3
Flask-OpenTracing==1.1.0
Jinja2==3.1.2
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
Werkzeug==2.3.8
Pillow==10.1.0
numpy
scipy
matplotlib==3.8.2
matplotlib-inline==0.1.3
squarify==0.4.3
seaborn==0.13.1